API_rest_SI/requirements.txt
flask
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
python-dotenv
gunicorn
pandas
mysql-connector-python